chore: move command metadata classes
