+++ packages/shorebird_cli/lib/src/metadata/build_environment_metadata.dart
