+++ packages/shorebird_cli/lib/src/metadata/metadata.dart
@@ -1,0 +1,3 @@
+export 'build_environment_metadata.dart';
+export 'create_patch_metadata.dart';
+export 'update_release_metadata.dart';
