+++ packages/shorebird_cli/pubspec.yaml
@@ -17,6 +17,7 @@ dependencies:
   cli_util: ^0.4.1
   collection: ^1.18.0
   crypto: ^3.0.5
+  equatable: ^2.0.5
   googleapis_auth: ^1.6.0
   http: ^1.2.1
   intl: ^0.19.0
