+++ packages/shorebird_cli/test/src/commands/release/releaser_test.dart
@@ -1,9 +1,9 @@
 import 'dart:io';
 
 import 'package:shorebird_cli/src/commands/release/releaser.dart';
+import 'package:shorebird_cli/src/metadata/metadata.dart';
 import 'package:shorebird_cli/src/release_type.dart';
 import 'package:shorebird_code_push_protocol/src/models/release.dart';
-import 'package:shorebird_code_push_protocol/src/models/update_release_metadata.dart';
 import 'package:test/test.dart';
 
 import '../../mocks.dart';
