+++ packages/shorebird_cli/test/src/metadata/create_patch_metadata_test.dart
@@ -1,13 +1,17 @@
+import 'package:shorebird_cli/src/metadata/metadata.dart';
 import 'package:shorebird_code_push_protocol/shorebird_code_push_protocol.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group(UpdateReleaseMetadata, () {
+  group(CreatePatchMetadata, () {
     test('can be (de)serialized', () {
-      const metadata = UpdateReleaseMetadata(
+      const metadata = CreatePatchMetadata(
         releasePlatform: ReleasePlatform.android,
-        flutterVersionOverride: '1.2.3',
-        generatedApks: false,
+        usedIgnoreAssetChangesFlag: false,
+        hasAssetChanges: false,
+        usedIgnoreNativeChangesFlag: false,
+        hasNativeChanges: false,
+        linkPercentage: 99.9,
         environment: BuildEnvironmentMetadata(
           flutterRevision: '853d13d954df3b6e9c2f07b72062f33c52a9a64b',
           operatingSystem: 'macos',
@@ -17,17 +21,20 @@ void main() {
         ),
       );
       expect(
-        UpdateReleaseMetadata.fromJson(metadata.toJson()).toJson(),
+        CreatePatchMetadata.fromJson(metadata.toJson()).toJson(),
         equals(metadata.toJson()),
       );
     });
 
     group('copyWith', () {
       test('creates a copy with the same fields', () {
-        const metadata = UpdateReleaseMetadata(
+        const metadata = CreatePatchMetadata(
           releasePlatform: ReleasePlatform.android,
-          flutterVersionOverride: '1.2.3',
-          generatedApks: false,
+          usedIgnoreAssetChangesFlag: false,
+          hasAssetChanges: false,
+          usedIgnoreNativeChangesFlag: false,
+          hasNativeChanges: false,
+          linkPercentage: 99.9,
           environment: BuildEnvironmentMetadata(
             flutterRevision: '853d13d954df3b6e9c2f07b72062f33c52a9a64b',
             operatingSystem: 'macos',
@@ -41,10 +48,13 @@ void main() {
       });
 
       test('creates a copy with the given fields replaced', () {
-        const metadata = UpdateReleaseMetadata(
+        const metadata = CreatePatchMetadata(
           releasePlatform: ReleasePlatform.android,
-          flutterVersionOverride: '1.2.3',
-          generatedApks: false,
+          usedIgnoreAssetChangesFlag: false,
+          hasAssetChanges: false,
+          usedIgnoreNativeChangesFlag: false,
+          hasNativeChanges: false,
+          linkPercentage: 99.9,
           environment: BuildEnvironmentMetadata(
             flutterRevision: '853d13d954df3b6e9c2f07b72062f33c52a9a64b',
             operatingSystem: 'macos',
@@ -56,8 +66,11 @@ void main() {
 
         final newMetadata = metadata.copyWith(
           releasePlatform: ReleasePlatform.ios,
-          flutterVersionOverride: '1.2.4',
-          generatedApks: true,
+          usedIgnoreAssetChangesFlag: true,
+          hasAssetChanges: true,
+          usedIgnoreNativeChangesFlag: true,
+          hasNativeChanges: true,
+          linkPercentage: 99.8,
           environment: const BuildEnvironmentMetadata(
             flutterRevision: 'asdf',
             operatingSystem: 'windows',
@@ -70,10 +83,13 @@ void main() {
         expect(
           newMetadata,
           equals(
-            const UpdateReleaseMetadata(
+            const CreatePatchMetadata(
               releasePlatform: ReleasePlatform.ios,
-              flutterVersionOverride: '1.2.4',
-              generatedApks: true,
+              usedIgnoreAssetChangesFlag: true,
+              hasAssetChanges: true,
+              usedIgnoreNativeChangesFlag: true,
+              hasNativeChanges: true,
+              linkPercentage: 99.8,
               environment: BuildEnvironmentMetadata(
                 flutterRevision: 'asdf',
                 operatingSystem: 'windows',
@@ -89,10 +105,13 @@ void main() {
 
     group('equatable', () {
       test('two metadatas with the same properties are equal', () {
-        const metadata = UpdateReleaseMetadata(
+        const metadata = CreatePatchMetadata(
           releasePlatform: ReleasePlatform.android,
-          flutterVersionOverride: '1.2.3',
-          generatedApks: false,
+          usedIgnoreAssetChangesFlag: false,
+          hasAssetChanges: false,
+          usedIgnoreNativeChangesFlag: false,
+          hasNativeChanges: false,
+          linkPercentage: 99.9,
           environment: BuildEnvironmentMetadata(
             flutterRevision: '853d13d954df3b6e9c2f07b72062f33c52a9a64b',
             operatingSystem: 'macos',
@@ -101,10 +120,13 @@ void main() {
             xcodeVersion: '15.0',
           ),
         );
-        const otherMetadata = UpdateReleaseMetadata(
+        const otherMetadata = CreatePatchMetadata(
           releasePlatform: ReleasePlatform.android,
-          flutterVersionOverride: '1.2.3',
-          generatedApks: false,
+          usedIgnoreAssetChangesFlag: false,
+          hasAssetChanges: false,
+          usedIgnoreNativeChangesFlag: false,
+          hasNativeChanges: false,
+          linkPercentage: 99.9,
           environment: BuildEnvironmentMetadata(
             flutterRevision: '853d13d954df3b6e9c2f07b72062f33c52a9a64b',
             operatingSystem: 'macos',
@@ -117,10 +139,12 @@ void main() {
       });
 
       test('two metadatas with different properties are not equal', () {
-        const metadata = UpdateReleaseMetadata(
+        const metadata = CreatePatchMetadata(
           releasePlatform: ReleasePlatform.android,
-          flutterVersionOverride: '1.2.3',
-          generatedApks: false,
+          usedIgnoreAssetChangesFlag: false,
+          hasAssetChanges: false,
+          usedIgnoreNativeChangesFlag: false,
+          hasNativeChanges: false,
           environment: BuildEnvironmentMetadata(
             flutterRevision: '853d13d954df3b6e9c2f07b72062f33c52a9a64b',
             operatingSystem: 'macos',
@@ -129,10 +153,12 @@ void main() {
             xcodeVersion: '15.0',
           ),
         );
-        const otherMetadata = UpdateReleaseMetadata(
-          releasePlatform: ReleasePlatform.android,
-          flutterVersionOverride: '1.2.4',
-          generatedApks: false,
+        const otherMetadata = CreatePatchMetadata(
+          releasePlatform: ReleasePlatform.ios,
+          usedIgnoreAssetChangesFlag: false,
+          hasAssetChanges: false,
+          usedIgnoreNativeChangesFlag: false,
+          hasNativeChanges: false,
           environment: BuildEnvironmentMetadata(
             flutterRevision: '853d13d954df3b6e9c2f07b72062f33c52a9a64b',
             operatingSystem: 'macos',
