+++ packages/shorebird_code_push_protocol/test/src/messages/create_patch/create_patch_request_test.dart
@@ -4,22 +4,9 @@ import 'package:test/test.dart';
 void main() {
   group(CreatePatchRequest, () {
     test('can be (de)serialized', () {
-      final request = CreatePatchRequest(
+      const request = CreatePatchRequest(
         releaseId: 1234,
-        metadata: const CreatePatchMetadata(
-          releasePlatform: ReleasePlatform.android,
-          usedIgnoreAssetChangesFlag: true,
-          usedIgnoreNativeChangesFlag: false,
-          hasAssetChanges: true,
-          hasNativeChanges: false,
-          linkPercentage: 99.9,
-          environment: BuildEnvironmentMetadata(
-            flutterRevision: '853d13d954df3b6e9c2f07b72062f33c52a9a64b',
-            operatingSystem: 'linux',
-            operatingSystemVersion: '1.0.0',
-            shorebirdVersion: '1.2.3',
-          ),
-        ).toJson(),
+        metadata: {'foo': 'bar'},
       );
       expect(
         CreatePatchRequest.fromJson(request.toJson()).toJson(),
@@ -28,9 +15,9 @@ void main() {
     });
 
     test('can be (de)serialized without metadata', () {
-      final request = CreatePatchRequest(
+      const request = CreatePatchRequest(
         releaseId: 1234,
-        metadata: CreatePatchMetadata.forTest().toJson(),
+        metadata: {'foo': 'bar'},
       );
       expect(
         CreatePatchRequest.fromJson(request.toJson()).toJson(),
